// src/pigeon/pigeon.interface.ts
export interface PigeonModuleOptions extends AedesOptions {
  port: number; // The port to listen on for MQTT connections
  transport: Transport; // The port to listen on for MQTT over WebSockets connections
  key?: Buffer; // The private key for secure connections
  cert?: Buffer; // The certificate for secure connections
  ca?: Buffer; // The certificate authority for secure connections
  rejectUnauthorized?: boolean; // Whether to reject unauthorized connections
  requestCert?: boolean; // Whether to request certificates from clients
}